test(api): strengthen sponsored resolution coverage
