+++ api/tests/test_sponsored_access.py
@@ -1,10 +1,8 @@
 import json
-import secrets
 from datetime import UTC, datetime
 
 import pytest
 
-from app.api.sponsored_access import resolve_sponsored_inference
 from app.llms.agents import sponsored_error_event
 from app.llms.model_access import (
     ModelAccessContext,
@@ -17,13 +15,11 @@ from app.llms.model_catalog_types import (
     ModelDescriptor,
 )
 from app.llms.provider_credentials import ProviderName
-from app.schemas.chat_credentials import ChatCredentialSecret
 from app.schemas.sponsored_access import (
     SPONSORED_ERROR_CODES,
     SafeErrorDetails,
     SponsoredQuotaSnapshot,
 )
-from app.utils.settings import Settings
 
 RESET = datetime(2026, 7, 19, tzinfo=UTC)
 
@@ -219,110 +215,3 @@ def test_sponsored_error_event_uses_error_event_shape_and_safe_reset() -> None:
         "free_tier_unavailable",
         "sponsored_request_in_progress",
     )
-
-
-def test_sponsored_resolution_prefers_user_credential_even_when_user_key_is_rejected() -> (
-    None
-):
-    user_credential = ChatCredentialSecret(
-        provider="openai",
-        api_key=secrets.token_urlsafe(),
-        base_url="https://user.example/v1",
-    )
-    settings = Settings(
-        SPONSORED_MODEL_ENABLED=True,
-        SPONSORED_MODEL_API_KEY="sponsored-key",
-        SPONSORED_MODEL_BASE_URL="HTTPS://Sponsored.EXAMPLE/v1/",
-        SPONSORED_MODEL_UPSTREAM_MODEL="upstream-luna",
-        SPONSORED_DAILY_GLOBAL_LIMIT=10,
-    )
-
-    resolved = resolve_sponsored_inference(
-        "gpt-5.6-luna",
-        user_credential,
-        settings,
-        user_credential_rejected=True,
-    )
-
-    assert resolved.credential == user_credential
-    assert resolved.sponsored is False
-    assert resolved.upstream_model is None
-
-
-def test_sponsored_resolution_builds_inference_only_sponsored_credential() -> None:
-    settings = Settings(
-        SPONSORED_MODEL_ENABLED=True,
-        SPONSORED_MODEL_API_KEY="sponsored-key",
-        SPONSORED_MODEL_BASE_URL="HTTPS://Sponsored.EXAMPLE/v1/",
-        SPONSORED_MODEL_UPSTREAM_MODEL="upstream-luna",
-        SPONSORED_MAX_OUTPUT_TOKENS=1024,
-        SPONSORED_DAILY_GLOBAL_LIMIT=10,
-    )
-
-    resolved = resolve_sponsored_inference("gpt-5.6-luna", None, settings)
-
-    assert resolved.credential == ChatCredentialSecret(
-        provider="openai",
-        api_key="sponsored-key",
-        base_url="https://sponsored.example/v1",
-    )
-    assert resolved.sponsored is True
-    assert resolved.upstream_model == "upstream-luna"
-    assert resolved.max_output_tokens == 1024
-
-
-def test_sponsored_resolution_does_not_sponsor_a_rejected_user_credential() -> None:
-    settings = Settings(
-        SPONSORED_MODEL_ENABLED=True,
-        SPONSORED_MODEL_API_KEY="sponsored-key",
-        SPONSORED_DAILY_GLOBAL_LIMIT=10,
-    )
-
-    resolved = resolve_sponsored_inference(
-        "gpt-5.6-luna",
-        None,
-        settings,
-        user_credential_rejected=True,
-    )
-
-    assert resolved.credential is None
-    assert resolved.sponsored is False
-
-
-def test_sponsored_resolution_targets_only_configured_model_id() -> None:
-    user_credential = ChatCredentialSecret(provider="google", api_key="user-key")
-    settings = Settings(
-        SPONSORED_MODEL_ENABLED=True,
-        SPONSORED_MODEL_ID="gpt-5.6-luna",
-        SPONSORED_MODEL_PROVIDER="openai",
-        SPONSORED_MODEL_API_KEY="sponsored-key",
-        SPONSORED_DAILY_GLOBAL_LIMIT=10,
-    )
-
-    resolved = resolve_sponsored_inference(
-        "gemini-3.5-flash",
-        user_credential,
-        settings,
-    )
-
-    assert resolved.credential == user_credential
-    assert resolved.sponsored is False
-    assert resolved.upstream_model is None
-
-
-def test_sponsored_resolution_uses_configured_provider_for_credential_secret() -> None:
-    settings = Settings(
-        SPONSORED_MODEL_ENABLED=True,
-        SPONSORED_MODEL_ID="gemini-3.5-flash",
-        SPONSORED_MODEL_PROVIDER="google",
-        SPONSORED_MODEL_API_KEY="sponsored-key",
-        SPONSORED_DAILY_GLOBAL_LIMIT=10,
-    )
-
-    resolved = resolve_sponsored_inference("gemini-3.5-flash", None, settings)
-
-    assert resolved.credential == ChatCredentialSecret(
-        provider="google",
-        api_key="sponsored-key",
-    )
-    assert resolved.sponsored is True
